requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain_huggingface
langchain_google_genai 
faiss-cpu
sentence-transformers
youtube-transcript-api
tiktoken 
python-dotenv 
streamlit
requests
